requirements.txt
nextcord
python-dotenv
selenium